corrigindo exclusao de armazem

+++ require/trans/deleteArmazem.php
@@ -4,9 +4,15 @@
 
   use Models\Armazem;
   $armazem    = new Armazem();
+  use Models\Estoque;
+  $estoque    = new Estoque();
 
-  $armazem->find($_POST['id']);
-  $armazem->delete();
-
-  echo "ok";
+  if(empty($estoque->getEstoqueByArmazem($_POST['id']))){
+    $armazem->find($_POST['id']);
+    $armazem->delete();
+    echo "ok";
+  } else {
+    echo "Não é possível remover um armazém que possui estoques.";
+  }
+  
 ?>
